third_party/WebKit/LayoutTests/fast/dom/document-set-title-mutations-expected.txt

Test for mutations to childList when setting document.title.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mutations.length is 1
PASS mutations[0].type is "childList"
PASS mutations[0].addedNodes.length is 1
PASS mutations[0].removedNodes.length is 1
PASS successfullyParsed is true

TEST COMPLETE
